requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.33
python-dotenv==1.0.0
werkzeug==2.3.7
Flask-WTF==1.2.1
email_validator==2.1.0.post1
